requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
frozenlist==1.3.3
idna==3.4
isort==5.12.0
magic-filter==1.0.9
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.7
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.2
yarl==1.9.2
